+++ noodles-cram/CHANGELOG.md
@@ -8,6 +8,10 @@
 
     This allows the header reader to be `Send`.
 
+### Fixed
+
+  * cram/record: Return empty data when no header is set.
+
 [#329]: https://github.com/zaeleus/noodles/issues/329
 
 ## 0.78.0 - 2025-02-17